+++ CHANGES.rst
@@ -1,5 +1,11 @@
 Changelog
 =========
+v0.2.0 (2014-10-13)
+- Renamed project to Mopidy-AudioAddict
+- Added support for the entire AudioAddict network. The extension
+  now plays DI.FM, RadioTunes, RockRadio, JazzRadio and FrescaRadio
+- No need to figure out your API key anymore, just enter your username
+  and password in the Mopidy configuration
 
 v0.1.0 (2014-08-12)
 -------------------